pyproject.toml
[tool.poetry]
name = "audiorecorder"
version = "0.1.0"
description = ""
authors = ["kayeungadrian-tam <tamkayeung.adrian@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyAudio = "^0.2.14"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
